requirements.txt
albumentations==1.3.1
hydra-core==1.3.2
numpy==1.25.1
omegaconf==2.3.0
opencv-python==4.8.0.74
phenobench==0.1.0
Pillow==10.0.0
torch==2.0.1
-e .